hw3/requirements.txt
gradio>=3.0
requests>=2.0
python-dotenv>=0.19
